refactor(header): get shows from useData
