+++ packages/mirror-tv-next/components/layout/header/header-nav.tsx
@@ -1,17 +1,15 @@
 import type { Category } from '~/graphql/query/category'
-import type { Show } from '~/graphql/query/shows'
 import styles from './_styles/header-nav.module.scss'
 import NavItems from './nav-items'
 
 type HeaderNavProps = {
   categories: Category[]
-  shows: Show[]
 }
 
-export default function HeaderNav({ categories, shows }: HeaderNavProps) {
+export default function HeaderNav({ categories }: HeaderNavProps) {
   return (
     <ul className={styles.ul}>
-      <NavItems categories={categories} shows={shows} />
+      <NavItems categories={categories} />
     </ul>
   )
 }
