+++ packages/mirror-tv-next/components/layout/header/mobile-header/mobile-nav.tsx
@@ -1,24 +1,17 @@
 import Image from 'next/image'
 import Link from 'next/link'
 import type { Category } from '~/graphql/query/category'
-import type { Show } from '~/graphql/query/shows'
 import type { Sponsor } from '~/graphql/query/sponsors'
-
 import styles from './_styles/mobile-nav.module.scss'
 import SideMenu from './side-menu'
 import MobileSearchBar from './mobile-search-bar'
 
 type MobileNavProps = {
   categories: Category[]
-  shows: Show[]
   sponsors: Sponsor[]
 }
 
-export default function MobileNav({
-  categories,
-  shows,
-  sponsors,
-}: MobileNavProps) {
+export default function MobileNav({ categories, sponsors }: MobileNavProps) {
   return (
     <div className={styles.navWrapper}>
       <div className={styles.logo}>
@@ -33,7 +26,7 @@ export default function MobileNav({
         </Link>
       </div>
       <section className={styles.sideBtnWrapper}>
-        <SideMenu categories={categories} shows={shows} sponsors={sponsors} />
+        <SideMenu categories={categories} sponsors={sponsors} />
         <MobileSearchBar />
       </section>
     </div>
